Check java before installing for testing

+++ lsst-dm-ci/run_tests.sh
@@ -23,6 +23,7 @@
 
 # Launched by LSST Jenkins-CI
 
+# stop execution on error
 set -e
 
 # setup lsst stack
@@ -30,7 +31,12 @@ source /opt/lsst/software/stack/loadLSST.bash
 setup lsst_distrib
 
 # install Java JRE for sodalint tests
-yum -y install java-11-openjdk
+if [ -f /usr/bin/java ]
+then
+  printf "Java already installed."
+else
+  yum -y install java-11-openjdk
+fi
 
 # setup imgserv
 cd /app
